requirements.txt
openai
python-dotenv
python-telegram-bot[job-queue]
pydub
